+++ crates/ruma-client-api/src/backup.rs
@@ -46,13 +46,31 @@ impl RoomKeyBackup {
 pub enum BackupAlgorithm {
     /// `m.megolm_backup.v1.curve25519-aes-sha2` backup algorithm.
     #[serde(rename = "m.megolm_backup.v1.curve25519-aes-sha2")]
-    MegolmBackupV1Curve25519AesSha2 {
-        /// The curve25519 public key used to encrypt the backups, encoded in unpadded base64.
-        public_key: Base64,
+    MegolmBackupV1Curve25519AesSha2(MegolmBackupV1Curve25519AesSha2AuthData),
+}
 
-        /// Signatures of the auth_data as Signed JSON.
-        signatures: CrossSigningOrDeviceSignatures,
-    },
+impl From<MegolmBackupV1Curve25519AesSha2AuthData> for BackupAlgorithm {
+    fn from(value: MegolmBackupV1Curve25519AesSha2AuthData) -> Self {
+        Self::MegolmBackupV1Curve25519AesSha2(value)
+    }
+}
+
+/// The data for the `m.megolm_backup.v1.curve25519-aes-sha2` backup algorithm.
+#[derive(Clone, Debug, Serialize, Deserialize)]
+#[cfg_attr(not(ruma_unstable_exhaustive_types), non_exhaustive)]
+pub struct MegolmBackupV1Curve25519AesSha2AuthData {
+    /// The curve25519 public key used to encrypt the backups, encoded in unpadded base64.
+    pub public_key: Base64,
+
+    /// Signatures of the auth_data as Signed JSON.
+    pub signatures: CrossSigningOrDeviceSignatures,
+}
+
+impl MegolmBackupV1Curve25519AesSha2AuthData {
+    /// Construct a new `MegolmBackupV1Curve25519AesSha2BackupAlgorithm` using the given public key.
+    pub fn new(public_key: Base64) -> Self {
+        Self { public_key, signatures: Default::default() }
+    }
 }
 
 /// Information about the backup key.
@@ -141,5 +159,57 @@ impl From<EncryptedSessionDataInit> for EncryptedSessionData {
     fn from(init: EncryptedSessionDataInit) -> Self {
         let EncryptedSessionDataInit { ephemeral, ciphertext, mac } = init;
         Self { ephemeral, ciphertext, mac }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use assert_matches2::assert_matches;
+    use ruma_common::{
+        SigningKeyAlgorithm, SigningKeyId, canonical_json::assert_to_canonical_json_eq,
+        owned_user_id, serde::Base64,
+    };
+    use serde_json::{from_value as from_json_value, json};
+
+    use super::{BackupAlgorithm, MegolmBackupV1Curve25519AesSha2AuthData};
+
+    #[test]
+    fn megolm_v1_backup_algorithm_serialize_roundtrip() {
+        let json = json!({
+            "algorithm": "m.megolm_backup.v1.curve25519-aes-sha2",
+            "auth_data": {
+                "public_key": "YWJjZGVm",
+                "signatures": {
+                    "@alice:example.org": {
+                        "ed25519:DEVICEID": "signature",
+                    },
+                },
+            },
+        });
+
+        let mut backup_algorithm =
+            MegolmBackupV1Curve25519AesSha2AuthData::new(Base64::new(b"abcdef".to_vec()));
+        backup_algorithm.signatures.insert_signature(
+            owned_user_id!("@alice:example.org"),
+            SigningKeyId::from_parts(SigningKeyAlgorithm::Ed25519, "DEVICEID".into()),
+            "signature".to_owned(),
+        );
+        assert_to_canonical_json_eq!(BackupAlgorithm::from(backup_algorithm), json.clone());
+
+        assert_matches!(
+            from_json_value(json),
+            Ok(BackupAlgorithm::MegolmBackupV1Curve25519AesSha2(auth_data))
+        );
+        assert_eq!(auth_data.public_key.as_bytes(), b"abcdef");
+        assert_matches!(
+            auth_data.signatures.get(&owned_user_id!("@alice:example.org")),
+            Some(user_signatures)
+        );
+
+        let mut user_signatures_iter = user_signatures.iter();
+        assert_matches!(user_signatures_iter.next(), Some((key_id, signature)));
+        assert_eq!(key_id, "ed25519:DEVICEID");
+        assert_eq!(signature, "signature");
+        assert_matches!(user_signatures_iter.next(), None);
     }
 }